// src/types/loan.types.ts
// ─── Local UI types ────────────────────────────────────────────────────────────
export type ActiveSection = "ASSET" | "LEGAL" | "SERVICING" | null;

export type DataIndex =
  | "name"
  | "Sponsor"
  | "Created"
  | "CreatedByTitle"
  | "Modified";

export type ModalType = "loan" | "folder" | "files" | "folderUpload" | null;